# openapi/climate-hub-v1.yaml
openapi: "3.0.3"
info:
  title: Climate Hub API — Need Engine Closed Loop
  version: "1.1.0"
  description: |
    REST API локальной платформы управления микроклиматом здания.
    Реализован замкнутый контур Need Engine: Environment → Need → Capability →
    Device → Command → Effect → Satisfaction.
servers:
  - url: /api/v1
    description: Climate Hub API
paths:
...
  /buildings/{buildingId}:
    get:
      operationId: getBuilding
      summary: Информация о здании
      tags: [Buildings]
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Данные здания
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildingSummary"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /rooms/{roomId}:
    get:
      operationId: getRoom
      summary: Информация о помещении
      tags: [Rooms]
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Данные помещения
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomSummary"
        "404":
          $ref: "#/components/responses/NotFound"

  # ────────────────────── Policy ──────────────────────
...
  /commands:
    post:
      operationId: createCommand
      summary: Создать команду
      tags: [Commands]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommandRequest"
      responses:
        "201":
          description: Команда создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"
    get:
      operationId: listCommands
      summary: Список команд
      tags: [Commands]
      responses:
        "200":
          description: Список команд
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommandResponse"
  /commands/{commandId}:
    get:
      operationId: getCommand
      summary: Информация о команде
      tags: [Commands]
      parameters:
        - name: commandId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Данные команды
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /commands/{commandId}/cancel:
    post:
      operationId: cancelCommand
      summary: Отменить команду
      tags: [Commands]
      parameters:
        - name: commandId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Команда отменена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"

  # ────────────────────── Needs ──────────────────────
  /needs:
    get:
      operationId: listNeeds
      summary: Список активных потребностей
      tags: [Needs]
      parameters:
        - name: roomId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NeedType"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NeedStatus"
        - name: severity
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NeedSeverity"
        - name: controlMode
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ControlMode"
        - name: activeOnly
          in: query
          required: false
          schema:
            type: boolean
        - name: blockedOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Список потребностей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NeedResponse"
  /needs/{needId}:
    get:
      operationId: getNeed
      summary: Детали потребности
      tags: [Needs]
      parameters:
        - name: needId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Детали потребности
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeedDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /needs/rooms/{roomId}:
    get:
      operationId: getNeedsByRoom
      summary: Потребности помещения
      tags: [Needs]
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Потребности помещения
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NeedResponse"
  /needs/buildings/{buildingId}:
    get:
      operationId: getNeedsByBuilding
      summary: Потребности здания
      tags: [Needs]
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Потребности здания
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NeedResponse"
  /needs/rooms/{roomId}/evaluate:
    post:
      operationId: evaluateRoom
      summary: Запустить оценку потребностей помещения
      tags: [Needs]
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Оценка выполнена
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: evaluated
  /needs/{needId}/evaluate:
    post:
      operationId: evaluateNeed
      summary: Запустить планирование для потребности
      tags: [Needs]
      parameters:
        - name: needId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Планирование выполнено
  /needs/{needId}/execute:
    post:
      operationId: executeNeed
      summary: Выполнить потребность (создать команду)
      tags: [Needs]
      parameters:
        - name: needId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Команда создана
        "400":
          description: Некорректный запрос
        "409":
          description: Уже есть активная команда
  /needs/{needId}/cancel:
    post:
      operationId: cancelNeed
      summary: Отменить потребность
      tags: [Needs]
      parameters:
        - name: needId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Потребность отменена
  /needs/{needId}/unblock:
    post:
      operationId: unblockNeed
      summary: Разблокировать потребность
      tags: [Needs]
      parameters:
        - name: needId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Потребность разблокирована
  /needs/{needId}/evaluations:
    get:
      operationId: getNeedEvaluations
      summary: История оценок потребности
      tags: [Needs]
      parameters:
        - name: needId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: История оценок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NeedEvaluationResponse"

  # ────────────────────── SSE Events ──────────────────────
...
components:
  schemas:
...
    ControlMode:
      type: string
      enum: [disabled, monitorOnly, manual, automatic]

    # ── Needs ──
    NeedType:
      type: string
      enum: [temperatureHeating, temperatureCooling, humidityIncrease, humidityDecrease, co2Reduction, illuminanceIncrease, illuminanceDecrease]
    NeedStatus:
      type: string
      enum: [detected, planning, planned, executing, waitingForEffect, satisfied, blocked, cancelled, expired]
    NeedSeverity:
      type: string
      enum: [low, medium, high, critical]
    NeedResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        buildingId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/NeedType"
        status:
          $ref: "#/components/schemas/NeedStatus"
        severity:
          $ref: "#/components/schemas/NeedSeverity"
        controlMode:
          $ref: "#/components/schemas/ControlMode"
        currentValue:
          type: number
          nullable: true
        deviation:
          type: number
        desiredMin:
          type: number
        desiredMax:
          type: number
        desiredPreferred:
          type: number
        sourceParameterCode:
          type: string
          nullable: true
        selectedCapabilityCode:
          type: string
          nullable: true
        selectedDeviceId:
          type: string
          format: uuid
          nullable: true
        activeCommandId:
          type: string
          format: uuid
          nullable: true
        planningFailureCode:
          type: string
          nullable: true
        violationSince:
          type: string
          format: date-time
          nullable: true
        stableSince:
          type: string
          format: date-time
          nullable: true
        cooldownUntil:
          type: string
          format: date-time
          nullable: true
        effectEvaluationDueAt:
          type: string
          format: date-time
          nullable: true
        lastEvaluationAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer
    NeedDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        buildingId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/NeedType"
        status:
          $ref: "#/components/schemas/NeedStatus"
        severity:
          $ref: "#/components/schemas/NeedSeverity"
        controlMode:
          $ref: "#/components/schemas/ControlMode"
        currentValue:
          type: number
          nullable: true
        desiredRange:
          type: object
          properties:
            minimum:
              type: number
            maximum:
              type: number
            preferred:
              type: number
        deviation:
          type: number
        sourceParameter:
          type: string
          nullable: true
        sourceMeasuredAt:
          type: string
          format: date-time
          nullable: true
        recommendation:
          type: object
          properties:
            capabilityCode:
              type: string
            operation:
              type: string
            requestedValue:
              type: number
            unit:
              type: string
            deviceId:
              type: string
              format: uuid
          nullable: true
        activeCommandId:
          type: string
          format: uuid
          nullable: true
        planningFailureCode:
          type: string
          nullable: true
        violationSince:
          type: string
          format: date-time
          nullable: true
        stableSince:
          type: string
          format: date-time
          nullable: true
        cooldownUntil:
          type: string
          format: date-time
          nullable: true
        effectEvaluationDueAt:
          type: string
          format: date-time
          nullable: true
        detectedAt:
          type: string
          format: date-time
        lastEvaluationAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer
    NeedEvaluationResponse:
      type: object
      properties:
        id:
          type: integer
        needId:
          type: string
          format: uuid
        trigger:
          type: string
        previousStatus:
          type: string
          nullable: true
        newStatus:
          type: string
          nullable: true
        outcome:
          type: string
          nullable: true
        failureCode:
          type: string
          nullable: true
        commandId:
          type: string
          format: uuid
          nullable: true
        evaluatedAt:
          type: string
          format: date-time
        correlationId:
          type: string
          nullable: true
...
    CreateCommandRequest:
      type: object
      properties:
        buildingId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        capabilityCode:
          type: string
        operation:
          type: string
        parametersJson:
          type: string
        priority:
          type: string
        createdBy:
          type: string
      required: [buildingId, deviceId, capabilityCode, operation]
    CommandResponse:
      type: object
      properties:
        commandId:
          type: string
          format: uuid
        status:
          type: string
        buildingId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
          nullable: true
        deviceId:
          type: string
          format: uuid
        capabilityCode:
          type: string
        operation:
          type: string
        parameters:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer

    # ── Buildings ──
    BuildingSummary:
      type: object
      required:
        - id
        - name
        - status
        - floorsCount
        - roomsCount
        - roomsWithWarnings
        - devicesOnline
        - devicesOffline
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [normal, warning, critical, offline, unknown]
        floorsCount:
          type: integer
        roomsCount:
          type: integer
        roomsWithWarnings:
          type: integer
        devicesOnline:
          type: integer
        devicesOffline:
          type: integer
        updatedAt:
          type: string
          format: date-time
...
    RoomSummary:
      type: object
      required:
        - id
        - floorId
        - name
        - status
        - environment
        - activeWarningsCount
        - devicesOnline
        - devicesTotal
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        floorId:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [normal, warning, critical, offline, stale, unknown]
        environment:
          $ref: "#/components/schemas/RoomEnvironmentSummary"
        activeWarningsCount:
          type: integer
        devicesOnline:
          type: integer
        devicesTotal:
          type: integer
        updatedAt:
          type: string
          format: date-time
    RoomEnvironmentSummary:
      type: object
      properties:
        temperatureC:
          type: number
          nullable: true
        relativeHumidityPct:
          type: number
          nullable: true
        co2Ppm:
          type: number
          nullable: true
        quality:
          $ref: "#/components/schemas/MeasurementQuality"
    MeasurementQuality:
      type: string
      enum: [valid, estimated, stale, outOfRange, sensorError, rejected, unknown]
...
    ApiProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errorCode:
          type: string
        traceId:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiProblemDetails"
...